Open confirm dialog on click

+++ src/components/ChatPreview.tsx
@@ -6,6 +6,7 @@ import clsx from "clsx";
 import { Pencil, X } from "lucide-react";
 import Link from "next/link";
 import React from "react";
+import { ConfirmDialog } from "./ConfirmDialog";
 
 interface ChatPreviewProps extends Chat {
   className?: string;
@@ -22,26 +23,14 @@ export const ChatPreview = ({
   const { messages, removeAllMessages } = useMessages();
 
   const handleDeleteChat = async (id: string) => {
-    const deleteChatConfirm = confirm("Do you really want to delete the chat?");
-    if (deleteChatConfirm) {
-      const deletedChatId = await deleteChat(id);
-      removeChat(deletedChatId);
-      setActiveChat(chats[0].id);
-    }
+    const deletedChatId = await deleteChat(id);
+    removeChat(deletedChatId);
+    setActiveChat(chats[0].id);
   };
 
   const handleChangeChat = (id: string) => {
-    if (messages.length > 1) {
-      const closeChat = confirm(
-        "Do you want to close current chat? Chat history will be removed."
-      );
-      if (closeChat) {
-        removeAllMessages();
-        setActiveChat(id);
-      }
-    } else {
-      setActiveChat(id);
-    }
+    removeAllMessages();
+    setActiveChat(id);
   };
 
   return (
@@ -50,16 +39,29 @@ export const ChatPreview = ({
         "relative group p-2 hover:bg-slate-300 hover:cursor-pointer transition-all",
         className
       )}
-      onClick={() => handleChangeChat(id)}
     >
-      <a
+      {messages.length > 1 ? (
+        <ConfirmDialog
+          text="Do you want to close current chat? Chat history will be removed."
+          confirmCallback={() => handleChangeChat(id)}
+        >
+          <div className="absolute top-0 left-0 w-full h-full" />
+        </ConfirmDialog>
+      ) : (
+        <div
+          className="absolute top-0 left-0 w-full h-full"
+          onClick={() => handleChangeChat(id)}
+        />
+      )}
+
+      <ConfirmDialog
         className="absolute top-2 right-2 hidden group-hover:block hover:text-white"
-        title="Delete Chat"
-        href="#"
-        onClick={() => handleDeleteChat(id)}
+        triggerTitle="Delete chat"
+        text="Do you really want to delete the chat?"
+        confirmCallback={() => handleDeleteChat(id)}
       >
         <X />
-      </a>
+      </ConfirmDialog>
       <a
         className="absolute top-2 right-10 hidden group-hover:block hover:text-white"
         title="Edit Chat"
